requirements.txt
python-dotenv~=1.1.0
python-telegram-bot~=22.1
google-genai~=1.16.1
google-auth~=2.40.1